Not necessary to show percent of volume with 5 numbers
svn path=/trunk/KDE/kdemultimedia/kscd/; revision=858857

+++ gui/panel.cpp
@@ -348,6 +348,6 @@ void Panel::setVolumeDisplay(qreal volume)
 {
 	timerVolume = 20;
 	QString s;
-	s = QString("<p align=right>%1</p>").arg(volume);
+	s = QString("<p align=right>%1</p>").arg((int)volume);
 	volumeDisplay->setText(s);
 }
